renamed and moved EvictorActor.Config
